Added a comment related to audio
- Added a comment for audio in N_SceneManager:
When loading an scene, all audios in audiomanager should be deleted to upload the new audios from the new scene. AudioManager updates the Wwise's transform of the audios (Wwise treats audios as gameobjects, so each audio has rotation and position). Right now, this audios are not being deleted so it should be checked for next assignment